+++ NEWS.md
@@ -3,6 +3,7 @@
 * simplified net_diag report by using only basic tools; works with html and pdf
 * better map for report_station
 * general package tidying
+* station functions check and return an error if report packages aren't installed.
 * removed the following functions: check_stn_gap.R, network_correlation.R
 
 # hydrolook 0.1.8